+++ R/AzureStor.R
@@ -12,6 +12,7 @@ globalVariables(c("self", "pool"), "AzureStor")
     options(azure_storage_api_version="2018-03-28")
     options(azure_adls_api_version="2018-06-17")
     options(azure_dl_progress_bar=TRUE)
+    options(azure_storage_retries=10)
 
     # all methods extending classes in external package must be run from .onLoad
     add_methods()